// src/swagger/model/transaction.ts
/**
 * fms
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Transaction { 
    processingDateTime?: Date;
    transactionDateTime?: Date;
    transactionAmount?: number;
    transactionId?: number;
    hexTagId?: string;
    vehicleNumber?: string;
    laneCode?: string;
    plazaCode?: string;
    transactionStatus?: string;
    transactionReferenceNumber?: string;
    plazaName?: string;
}